// src/abi/OrderbookAbi.d.ts
export enum AssetTypeOutput { Base = 'Base', Quote = 'Quote' };
export enum AuthErrorInput { Unauthorized = 'Unauthorized' };
export enum AuthErrorOutput { Unauthorized = 'Unauthorized' };
export type IdentityInput = Enum<{ Address: AddressInput, ContractId: ContractIdInput }>;
export type IdentityOutput = Enum<{ Address: AddressOutput, ContractId: ContractIdOutput }>;
// ...
export enum OrderTypeOutput { Buy = 'Buy', Sell = 'Sell' };
export enum ValueErrorInput { InvalidAmount = 'InvalidAmount', InvalidArrayLength = 'InvalidArrayLength' };
export enum ValueErrorOutput { InvalidAmount = 'InvalidAmount', InvalidArrayLength = 'InvalidArrayLength' };

export type AccountInput = { liquid: BalanceInput, locked: BalanceInput };
export type AccountOutput = { liquid: BalanceOutput, locked: BalanceOutput };
export type AddressInput = { bits: string };
export type AddressOutput = AddressInput;
export type AssetIdInput = { bits: string };
export type AssetIdOutput = AssetIdInput;
export type BalanceInput = { base: BigNumberish, quote: BigNumberish };
export type BalanceOutput = { base: BN, quote: BN };
export type CancelOrderEventInput = { order_id: string };
export type CancelOrderEventOutput = CancelOrderEventInput;
export type ContractIdInput = { bits: string };
export type ContractIdOutput = ContractIdInput;
export type DepositEventInput = { amount: BigNumberish, asset: AssetIdInput, user: IdentityInput };
export type DepositEventOutput = { amount: BN, asset: AssetIdOutput, user: IdentityOutput };
export type MatchOrderEventInput = { order_id: string, asset: AssetIdInput, order_matcher: IdentityInput, owner: IdentityInput, counterparty: IdentityInput, match_size: BigNumberish, match_price: BigNumberish };
export type MatchOrderEventOutput = { order_id: string, asset: AssetIdOutput, order_matcher: IdentityOutput, owner: IdentityOutput, counterparty: IdentityOutput, match_size: BN, match_price: BN };
export type OpenOrderEventInput = { amount: BigNumberish, asset: AssetIdInput, asset_type: AssetTypeInput, order_type: OrderTypeInput, order_id: string, price: BigNumberish, user: IdentityInput };
export type OpenOrderEventOutput = { amount: BN, asset: AssetIdOutput, asset_type: AssetTypeOutput, order_type: OrderTypeOutput, order_id: string, price: BN, user: IdentityOutput };
// ...
export type TradeOrderEventOutput = { base_sell_order_id: string, base_buy_order_id: string, order_matcher: IdentityOutput, trade_size: BN, trade_price: BN, block_height: number, tx_id: string };
export type WithdrawEventInput = { amount: BigNumberish, asset: AssetIdInput, user: IdentityInput };
export type WithdrawEventOutput = { amount: BN, asset: AssetIdOutput, user: IdentityOutput };